detect when test suite is run from within a virtualenv

+++ runtests.sh
@@ -1,14 +1,29 @@
 #!/bin/bash
+#
+# Run the Pyflame test suite.
+#
+# If invoked without arguments, this will make a best effort to run the test
+# suite against python2 and python3. If you would like to force the test suite
+# to run against a specific version of python, invoke with the python
+# interpreter names as arguments. These should be strings suitable for passing
+# to virtualenv -p.
 
 set -e
 
-ENVDIR="./test_env"
+# This script creates its own virtualenv and things can get weird if you try to
+# nest them.
+if [ -n "${VIRTUAL_ENV}" ]; then
+  echo "Do not invoke runtests.sh from within a virtualenv!"
+  exit 1
+fi
+
+ENVDIR="./.test_env"
 trap 'rm -rf ${ENVDIR}' EXIT
 
 # Run tests using pip; $1 = python version
 run_pip_tests() {
   rm -rf "${ENVDIR}"
-  virtualenv -q -p "$1" "${ENVDIR}"
+  virtualenv -q -p "$1" "${ENVDIR}" &>/dev/null
 
   . "${ENVDIR}/bin/activate"
   pip install -q pytest
@@ -20,20 +35,20 @@ run_pip_tests() {
 # Make a best effort to run the tests against some Python version.
 try_pip_tests() {
   if command -v "$1" &>/dev/null; then
-    echo -n "Running test suite against "
+    echo -n "Running test suite against interpreter "
     "$1" --version
     run_pip_tests "$1"
   fi
 }
 
-# RPM tests are not allowed to use a virtualenv.
+# RPM tests are not allowed to use virtualenv at all.
 run_rpm_tests() {
   py.test-2 tests/
   py.test-3 tests/
 }
 
 if [ $# -eq 0 ]; then
-  try_pip_tests python
+  try_pip_tests python2
   try_pip_tests python3
 elif [ "$1" = "rpm" ]; then
   run_rpm_tests
